requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-CORS==4.0.0
google-play-scraper==1.2.4
groq==0.4.1
hypothesis==6.92.1
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
langdetect==1.0.9
schedule==1.2.0
pytz==2024.1
